fix：修复字节溢出问题

+++ BSP/Sensor/Sensor.h
@@ -33,10 +33,10 @@ typedef struct {
  * @brief 压力传感器数据结构（纯数据）
  */
 typedef struct {
- float val;        /**< 转换后的压力值（单位：根据实际定义） */
- int32_t raw_val;  /**< 原始32位有符号值 */
- float filter_val;
- uint8_t id;       /**< 传感器ID */
+ int32_t val;        /**< 转换后的压力值（单位：根据实际定义） */
+ int32_t raw_val;    /**< 原始32位有符号值 */
+ int32_t filter_val;
+ uint8_t id;         /**< 传感器ID */
 } PressSensor;
 
 /**
